rename data to data_util
