+++ model_zoo/ernie-vil2.0/data_util.py
